refactor(redaction): F14 isolate workspace application policy behind ports
Preserve the existing IPC/job facade while an application service owns sessions, save/approval ordering and cleanup. Inject draft/approval storage, fingerprints and diagnostics; application code imports no Electron, filesystem adapters or runtime initialization. Keep F03 merge semantics unchanged.

11 application/concurrency/pending-review cases, focused ESLint, formatting and architecture budget passed locally before unrelated concurrent F17/SFX commits. Full Check still pending. F14 complete; next here F07 scaling. Preserve concurrent preview/mask/error/lifetime work. No master, version or release changes.
